drivers/smm_payload_interface: Add initial support for SMM payload
UEFI Secure Boot is a security feature that protects the chain-of-trust.
With proper implementation and configuration, it ensures that the user's
bootloader and OS start as desired, without malicious modification.

The protection of key variables is critical to Secure Boot, so, note
that the UEFI variable protocol provides the Variable Policy to
restrict write permissions, and variables have an `Attributes` field,
enabling protections such as hiding at runtime or "authenticating" write
requests.

Presently, coreboot's UefiPayload can support the Authenticated
Variables required by Secure Boot using SMMSTORE. Simply use this GUID
signature on the variable store to advertise support. However, this does
not offer the expected guarantees: Usually, when `gRT->SetVariable()` is
called, SMM is entered via EDK2 "SMM communication" for VariableSmm to
perform verification of the write request, then write SPI flash.

However, the SMMSTORE SMI handler can be initiated manually (outside of
the `gRT->SetVariable()` -> FaultTolerantWrite protocol ->
FirmwareVolumeBlock protocol == SMMSTORE). In this case, the variable
driver cannot verify a write request is permitted. In addition, the
variable driver - in this case, VariableRuntimeDxe - will be vulnerable
to malicious patching. Therefore, a more integrated solution is required
for Secure Boot.

The solution here is to initialise SMM in EDK2, following UefiPayload
work for slimbootloader. This approach has several challenges, to be
resolved as below:
- coreboot provides register definitions in CBMEM to avoid
  silicon-dependence in payload.
- Payloads are not on the S3 resume path. While TSEG is preserved,
  coreboot must perform SMM relocate and trigger a payload SMI, which
  will lock SMM using the SMRRs. These are privileged MSRs.

A minimal 4K communication region is shared with the payload, located
in SMRAM. CPU-specific SMBASEs are stored here, to be provided to common
code for relocation, as well as the payload's SMI number to trigger for
final SMM lockdown.

TEST= Successful with UefiPayload patches, tested on ADLRVP.
- Cold boot: Verify that payload initialises SMM and writes variables.
- S3 resume: Verify that SMI succeeds with debug log output.

Change-Id: I3dd505a154175b6da8929b1157723de1645ca8fa
Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>